+++ packages/okta-react/src/Security.js
@@ -10,19 +10,25 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import AuthService from './AuthService';
 import OktaContext from './OktaContext';
 
 const Security = (props) => { 
 
-  const [authService] = useState( props.authService || new AuthService(props) );
-  const [authState, setAuthState] = useState(authService.getAuthState());
+  const initialAuthService = useMemo( () => { 
+    // don't keep spawning new service instances if this component rerenders
+    return props.authService || new AuthService(props);
+  }, [ props ]);
 
+  const [authService] = useState( initialAuthService );
+  const [authState, setAuthState] = useState(authService.getAuthState());
+  
   useEffect( () => { 
     const unsub = authService.on('authStateChange', () => {
       setAuthState(authService.getAuthState());
     });
+
     authService.updateAuthState(); // Trigger an initial change event to make sure authState is latest
     return unsub;
   }, [authService]);