+++ packages/okta-react/CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Bug Fixes
 
 - `<LoginCallback>` now triggers only after `authState.isPending` is false, removing the problem of as error message from parsing the tokens from the url being cleared by the pending `authState` determination. See [#719](https://github.com/okta/okta-oidc-js/issues/719)
+- `<Security>` now memoizes if it creates an instance of `AuthService` so as to not create new instances on re-renders
 
 # 3.0.0